--- ToAdd.txt
@@ -1,48 +1,0 @@
-Should scrap:
-- 20 stray beads / 50 stray beads (10 stray beads is reasonable and can stay)
-- CB3 and PS3 honestly, CB2 and PS2 are on thin ice, def subject to discussion
-- increase solar energy to max
-- feed "all animals" in (x location) aint nobody know where all animals are, im sorry
-- bloom all the clovers in 100 YITP shinshu
-- defeat orochi from 100 YITP
-- get praise from sleepy hbs kiiiinda makes this too free, i bring alternatives in suggestions
-- obtain 100 demon fangs
-- clear nryo/kamui devil gate (ryo could stay? since its the fastest of the three, but nryo and especially kdg is......Uh.)
-- obtain all the stray beads in kamui/nryo
-- "get 13 brush techniques (no hidden ones)"
-- deliver a vase to the orochi guardian statue
-
-Should add:
-- bloom 10 cursed patches (iffy on the number, could be 5 or 7?)
-- deliver the acorn to sleepy
-- deliver the beehive to sleepy
-- deliver inaba to the animal lover
-- take susano to nagi's statue
-- obtain 40 demon fangs (iffy on the number)
-- use 15 bags of meat
-- use 15 bags of grass
-- use 15 bags of seeds
-- use 15 bags of fish
-- find 5 zodiac statues
-- race kai
-- deliver the prayer slips to rao
-- upgrade a weapon with gold dust
-- defeat two blockheads
-- defeat the cutters
-- get 3 hidden brush techniques (includes upgrades like PS2/CB2)
-- cut down a fruit from sakuyas tree (excluding the fruit from the start of the game)
-- do one of yoichi's quests
-- raid the treasury in jamba's room
-- clear 6 devil gates
-- eat a cherry cake prepared by mrs. orange
-- learn hardhead
-- get 10 brush techniques (including hidden ones / upgrades)
-- dig holes in the ground until money pops up
-- deliver a vase to three guardian statues
-- yeet a lockjaw key
-- do two digging minigames
-- get fed by three separate npcs
-- get whacked by three separate npcs
-- make five snow statues in kamui
-- race tobi three times
-- fish everywhere it is possible to fish at least once (benkei in sei-an counts as one despite his location change)
